Update Vector_GradientDescend.py
如果指定数据类型，计算出来的结果会与非向量化的结算结果不同，这因为数据类型不一样导致的

+++ Logistic/Vector_GradientDescend.py
@@ -13,7 +13,8 @@ def main():
     # X为输入训练集，有NUM_TRAIN个训练样本,每个样本有NUM_CHARATER个特征
     # Y是标签
     X = rdm.rand(NUM_CHARACTER, NUM_TRAIN)
-    Y = np.zeros(shape=[1, NUM_TRAIN], dtype=np.int)
+    Y = np.zeros(shape=[1, NUM_TRAIN])
+    # Y = np.zeros(shape=[1, NUM_TRAIN], dtype=np.int)   如果指定数据类型，计算出来的结果会与非向量化的结算结果不同，这因为数据类型不一样导致的
 
     # print(X.T.shape)
     i = 0
